+++ components/proxy/src/main/java/com/hotels/styx/proxy/ProxyConnectorFactory.java
@@ -1,5 +1,5 @@
 /*
-  Copyright (C) 2013-2018 Expedia Inc.
+  Copyright (C) 2013-2019 Expedia Inc.
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -29,8 +29,8 @@ import com.hotels.styx.server.netty.ServerConnectorFactory;
 import com.hotels.styx.server.netty.codec.NettyToStyxRequestDecoder;
 import com.hotels.styx.server.netty.connectors.HttpPipelineHandler;
 import com.hotels.styx.server.netty.connectors.ResponseEnhancer;
+import com.hotels.styx.server.netty.handlers.ChannelActivityEventConstrainer;
 import com.hotels.styx.server.netty.handlers.ChannelStatisticsHandler;
-import com.hotels.styx.server.netty.handlers.ConnectionCountHandler;
 import com.hotels.styx.server.netty.handlers.ExcessConnectionRejector;
 import com.hotels.styx.server.netty.handlers.RequestTimeoutHandler;
 import com.hotels.styx.server.track.CurrentRequestTracker;
@@ -148,9 +148,9 @@ class ProxyConnectorFactory implements ServerConnectorFactory {
 
             channel.pipeline()
                     .addLast("connection-throttler", excessConnectionRejector)
+                    .addLast("channel-activity-event-constrainer", new ChannelActivityEventConstrainer())
                     .addLast("idle-handler", new IdleStateHandler(serverConfig.requestTimeoutMillis(), 0, serverConfig.keepAliveTimeoutMillis(), MILLISECONDS))
                     .addLast("channel-stats", channelStatsHandler)
-                    .addLast("channel-connection-stats", new ConnectionCountHandler(metrics))
 
                     // Http Server Codec
                     .addLast("http-server-codec", new HttpServerCodec(serverConfig.maxInitialLineLength(), serverConfig.maxHeaderSize(), serverConfig.maxChunkSize(), true))
