+++ system-tests/e2e-suite/src/test/scala/com/hotels/styx/server/ServerConnectionsSpec.scala
@@ -1,5 +1,5 @@
 /*
-  Copyright (C) 2013-2018 Expedia Inc.
+  Copyright (C) 2013-2019 Expedia Inc.
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -17,7 +17,7 @@ package com.hotels.styx.server
 
 import java.nio.charset.StandardCharsets.UTF_8
 import java.util
-import java.util.concurrent.{CompletableFuture, ExecutionException}
+import java.util.concurrent.CompletableFuture
 import java.util.concurrent.TimeUnit.SECONDS
 
 import com.hotels.styx.StyxProxySpec
